work in progress update

+++ nRF5_SDK_17.0.2_d674dde/examples/peripheral/pwm_driver_ws2812b/Src/main.c
@@ -82,7 +82,7 @@
 /// \return    none
 int main( void )
 {
-   static uint8_t i;
+   static uint32_t i;
 
    WS2812B_init();
